terrasync-eartheye/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
Pillow==10.3.0
opencv-python==4.9.0.80
python-multipart==0.0.9
scikit-learn==1.4.2
matplotlib==3.8.4
mlflow==2.11.3
tqdm==4.66.2
sqlalchemy==2.0.29
